pyproject.toml
[tool.poetry]
name = "boa_console"
version = "0.1.0"
description = "A web3 utility package"
authors = ["Dylan Garrett"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
web3 = "^6.10.0"
eth-tester = "^0.9.1b1"
ipython = "^8.16.1"
inquirerpy = "^0.3.4"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
